requirements.txt
streamlit==1.51.0
zhipuai==2.1.5.20250725
python-dotenv==1.0.0
